changed isizes

+++ src/vertex_buffer.rs
@@ -130,9 +130,9 @@ impl<'a> GLByteBuffer<'a> {
         gl::CopyBufferSubData(
           gl::ARRAY_BUFFER,
           gl::ARRAY_BUFFER,
-          self.length,
-          i,
-          count,
+          self.length as i32,
+          i as i32,
+          count as i32,
         );
       }
     }
@@ -154,8 +154,8 @@ impl<'a> GLByteBuffer<'a> {
 
     gl::BufferSubData(
       gl::ARRAY_BUFFER,
-      idx,
-      count,
+      idx as i32,
+      count as i32,
       mem::transmute(vs)
     );
   }
